requirements.txt
streamlit>=1.37
feedparser>=6.0.11
rich>=13.7.1
typer>=0.12.5
pandas>=2.2.2
python-dateutil>=2.9.0.post0
pytz>=2024.1
